+++ structure/parser/object_parser_test.go
@@ -838,4 +838,17 @@ var _ = Describe("Object", func() {
 			Expect(base.Error()).ToNot(HaveOccurred())
 		})
 	})
+
+	Context("WithReferenceErrorReporter", func() {
+		var parser *structureParser.Object
+
+		BeforeEach(func() {
+			parser = structureParser.NewObjectParser(base, nil)
+			Expect(parser).ToNot(BeNil())
+		})
+
+		It("with reference returns not nil", func() {
+			Expect(parser.WithReferenceErrorReporter("reference")).ToNot(BeNil())
+		})
+	})
 })